Drop 'status' column from Ban + deprecate its use
The use of SET columns in MySQL is discouraged for a number of reasons:

- Not cross compatible with other database
- Modifying the SET column in the future is an expensive operation

In order to best support different databases in the future, the use of
the 'status' column for models has been deprecated. In order to check
for deleted objects, the 'is_deleted' column will become the new
convention for our models.

Any other states a model may have will follow this convention, e.g. a
model that is still a draft would use an 'is_draft' column.

The first model being converted will be the Ban model.